Wrote reverse all of a linked list test that will be built on
