+++ src/main/java/com/lperthel/datastructures/Solution.java
@@ -1,12 +1,7 @@
 package com.lperthel.datastructures;
-class Solution {
-    public ListNode reverseList(ListNode head) {
-    	if(head== null || head.next == null)
-return head;
-ListNode nextNode = head.next;
-ListNode tail = reverseList(nextNode);
-nextNode.next = head;
-head.next = null;
-return tail;
-    }
+
+public class Solution {
+    public ListNode reverseBetween(ListNode head, int left, int right) {
+     return   null;    
+     }
 }
